+++ demo/src/components/controls/number.tsx
@@ -46,8 +46,15 @@ export default function NumberControlView() {
           mask=""
           maskholder=""
         />
-        {/* todo implement format */}
-        <wup-num name="decimal" initValue="10.237" />
+        <wup-num
+          name="decimal"
+          initValue="10.237"
+          ref={(el) => {
+            if (el) {
+              el.$options.format = { maxDecimal: 2 };
+            }
+          }}
+        />
         <wup-num name="NumberRequired" validations="myRequiredValidations" />
         <wup-num name="WithMask" mask="0000-0000" />
         <wup-num